Consolidate DateTimeVariantTest.cs into DateTimeVariantTests.cs
Move all helper methods (xsql, DropStoredProcedure, DropTable, DropType,
GetSqlDbType, RunTest) and the ExceptionChecker delegates from the old
DateTimeVariantTest.cs static class into DateTimeVariantTests.cs. Each
test variation is now a separate [ConditionalTheory] test method directly
in the test class. Remove DateTimeVariantTest.cs and its csproj entry.
