+++ src/Microsoft.Data.SqlClient/tests/ManualTests/SQL/ParameterTest/DateTimeVariantTests.cs
@@ -4,42 +4,1116 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Text;
+using System.Data;
+using Microsoft.Data.SqlClient.Server;
 using Xunit;
 
 #nullable enable
 namespace Microsoft.Data.SqlClient.ManualTesting.Tests
 {
+    public enum TestVariations {
+        TestSimpleParameter_Type,
+        TestSimpleParameter_Variant,
+        TestSqlDataRecordParameterToTVP_Type,
+        TestSqlDataRecordParameterToTVP_Variant,
+        TestSqlDataReaderParameterToTVP_Type,
+        TestSqlDataReaderParameterToTVP_Variant,
+        TestSqlDataReader_TVP_Type,
+        TestSqlDataReader_TVP_Variant,
+        TestSimpleDataReader_Type,
+        TestSimpleDataReader_Variant,
+        SqlBulkCopySqlDataReader_Type,
+        SqlBulkCopySqlDataReader_Variant,
+        SqlBulkCopyDataTable_Type,
+        SqlBulkCopyDataTable_Variant,
+        SqlBulkCopyDataRow_Type,
+        SqlBulkCopyDataRow_Variant
+    };
+
+    public struct TestResult
+    {
+        public object Value { get; }
+        public string BaseTypeName { get; }
+
+        public TestResult(object value, string baseTypeName)
+        {
+            Value = value;
+            BaseTypeName = baseTypeName;
+        }
+    }
+
+    public delegate bool ExceptionChecker(Exception e, object paramValue);
+
     /// <summary>
     /// Tests for DateTime variant parameters with different date/time types.
-    /// These tests run independently with their own baseline comparison.
     /// </summary>
     public class DateTimeVariantTests
     {
-        private const string BaselineDirectory = "DateTimeVariant";
+        private static void RunTest(
+            TestVariations tag,
+            Func<object, string, string, TestResult> action,
+            object paramValue, 
+            string expectedBaseTypeName, 
+            string connStr, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            try
+            {
+                TestResult result = action(paramValue, expectedBaseTypeName, connStr);
 
-        /// <summary>
-        /// Parameterized test for DateTime variant parameters.
-        /// Each parameter combination runs 16 test methods and compares against 16 baseline files.
-        /// </summary>
+                expectedValueOverrides.TryGetValue(tag, out var expectedValueOverride);
+                expectedBaseTypeOverrides.TryGetValue(tag, out var expectedBaseTypeOverride);
+                
+                Assert.Equal(expectedValueOverride ?? paramValue, result.Value);
+                Assert.Equal(expectedBaseTypeOverride ?? expectedBaseTypeName, result.BaseTypeName);
+            }
+            catch (Exception e)
+            {
+                if (expectedExceptions.TryGetValue(tag, out var isExpectedException))
+                {
+                    Assert.True(isExpectedException(e, paramValue), e.Message);
+                }
+                else {
+                    Assert.Fail($"Unexpected exception was thrown for test variation {tag} with parameter value {paramValue}. Exception: {e}");
+                }
+            }
+        }
+
+        private static SqlDbType GetSqlDbType(string expectedBaseTypeName)
+        {
+            return expectedBaseTypeName.ToLowerInvariant() switch
+            {
+                "time" => SqlDbType.Time,
+                "date" => SqlDbType.Date,
+                "smalldatetime" => SqlDbType.SmallDateTime,
+                "datetime" => SqlDbType.DateTime,
+                "datetime2" => SqlDbType.DateTime2,
+                "datetimeoffset" => SqlDbType.DateTimeOffset,
+                _ => SqlDbType.Variant,
+            };
+        }
+
+        private static void xsql(SqlConnection conn, string sql)
+        {
+            using SqlCommand cmd = conn.CreateCommand();
+            cmd.CommandText = sql;
+            cmd.ExecuteNonQuery();
+        }
+
+        private static void DropStoredProcedure(SqlConnection conn, string procName)
+        {
+            xsql(conn, string.Format("if exists(select 1 from sys.procedures where name='{0}') begin drop proc {1} end", procName.Substring(1, procName.Length - 2), procName));
+        }
+
+        private static void DropTable(SqlConnection conn, string tableName)
+        {
+            xsql(conn, string.Format("if exists(select 1 from sys.tables where name='{0}') begin drop table {1} end", tableName.Substring(1, tableName.Length - 2), tableName));
+        }
+
+        private static void DropType(SqlConnection conn, string typeName)
+        {
+            xsql(conn, string.Format("if exists(select 1 from sys.types where name='{0}') begin drop type {1} end", typeName.Substring(1, typeName.Length - 2), typeName));
+        }
+
         [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
         [MemberData(nameof(GetParameterCombinations))]
-        public void DateTimeVariantParameterTest(
+        public void TestSimpleParameter_Type(
             object paramValue, 
-            string expectedTypeName, 
             string expectedBaseTypeName, 
-            Dictionary<TestVariations, ExceptionChecker>? expectedExceptions = null, 
-            Dictionary<TestVariations, object>? expectedValueOverrides = null,
-            Dictionary<TestVariations, object>? expectedBaseTypeOverrides = null)
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
         {
-            expectedExceptions ??= new Dictionary<TestVariations, ExceptionChecker>();
-            expectedValueOverrides ??= new Dictionary<TestVariations, object>();
-            expectedBaseTypeOverrides ??= new Dictionary<TestVariations, object>();
+            RunTest(
+                TestVariations.TestSimpleParameter_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string procName = DataTestUtility.GetLongName("paramProc1");
 
-            DateTimeVariantTest.SendInfo(
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropStoredProcedure(conn, procName);
+                        xsql(conn, string.Format("create proc {0} (@param {1}) as begin select @param, sql_variant_property(@param,'BaseType') as BaseType end;", procName, expectedBaseTypeName));
+
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = procName;
+                        cmd.CommandType = CommandType.StoredProcedure;
+                        SqlParameter p = cmd.Parameters.AddWithValue("@param", paramValue);
+                        cmd.Parameters[0].SqlDbType = GetSqlDbType(expectedBaseTypeName);
+                        using (SqlDataReader dr = cmd.ExecuteReader())
+                        {
+                            dr.Read();
+                            return new TestResult(dr[0], dr.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropStoredProcedure(conn, procName);
+                    }
+                },
                 paramValue, 
-                expectedTypeName, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSimpleParameter_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSimpleParameter_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string procName = DataTestUtility.GetLongName("paramProc2");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropStoredProcedure(conn, procName);
+                        xsql(conn, string.Format("create proc {0} (@param sql_variant) as begin select @param, sql_variant_property(@param,'BaseType') as BaseType end;", procName));
+
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = procName;
+                        cmd.CommandType = CommandType.StoredProcedure;
+                        SqlParameter p = cmd.Parameters.AddWithValue("@param", paramValue);
+                        cmd.Parameters[0].SqlDbType = SqlDbType.Variant;
+                        using (SqlDataReader dr = cmd.ExecuteReader())
+                        {
+                            dr.Read();
+                            return new TestResult(dr[0], dr.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropStoredProcedure(conn, procName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSqlDataRecordParameterToTVP_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSqlDataRecordParameterToTVP_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string tvpTypeName = DataTestUtility.GetLongName("tvpType");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropType(conn, tvpTypeName);
+                        xsql(conn, string.Format("create type dbo.{0} as table (f1 {1})", tvpTypeName, expectedBaseTypeName));
+
+                        // Send TVP using SqlMetaData.
+                        SqlMetaData[] metadata = new SqlMetaData[1];
+                        metadata[0] = new SqlMetaData("f1", GetSqlDbType(expectedBaseTypeName));
+                        SqlDataRecord[] record = new SqlDataRecord[1];
+                        record[0] = new SqlDataRecord(metadata);
+                        record[0].SetValue(0, paramValue);
+
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = "select f1, sql_variant_property(f1,'BaseType') as BaseType from @tvpParam";
+                        SqlParameter p = cmd.Parameters.AddWithValue("@tvpParam", record);
+                        p.SqlDbType = SqlDbType.Structured;
+                        p.TypeName = string.Format("dbo.{0}", tvpTypeName);
+                        using (SqlDataReader dr = cmd.ExecuteReader())
+                        {
+                            dr.Read();
+                            return new TestResult(dr[0], dr.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropType(conn, tvpTypeName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSqlDataRecordParameterToTVP_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSqlDataRecordParameterToTVP_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string tvpTypeName = DataTestUtility.GetLongName("tvpVariant");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropType(conn, tvpTypeName);
+                        xsql(conn, string.Format("create type dbo.{0} as table (f1 sql_variant)", tvpTypeName));
+
+                        // Send TVP using SqlMetaData.
+                        SqlMetaData[] metadata = new SqlMetaData[1];
+                        metadata[0] = new SqlMetaData("f1", SqlDbType.Variant);
+                        SqlDataRecord[] record = new SqlDataRecord[1];
+                        record[0] = new SqlDataRecord(metadata);
+                        record[0].SetValue(0, paramValue);
+
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = "select f1, sql_variant_property(f1,'BaseType') as BaseType from @tvpParam";
+                        SqlParameter p = cmd.Parameters.AddWithValue("@tvpParam", record);
+                        p.SqlDbType = SqlDbType.Structured;
+                        p.TypeName = string.Format("dbo.{0}", tvpTypeName);
+                        using (SqlDataReader dr = cmd.ExecuteReader())
+                        {
+                            dr.Read();
+                            return new TestResult(dr[0], dr.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropType(conn, tvpTypeName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSqlDataReaderParameterToTVP_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSqlDataReaderParameterToTVP_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string tvpTypeName = DataTestUtility.GetLongName("tvpType");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropType(conn, tvpTypeName);
+                        xsql(conn, string.Format("create type dbo.{0} as table (f1 {1})", tvpTypeName, expectedBaseTypeName));
+                        using (SqlConnection connInput = new(connStr))
+                        {
+                            connInput.Open();
+                            using (SqlCommand cmdInput = connInput.CreateCommand())
+                            {
+                                cmdInput.CommandText = "select @p1 as f1";
+                                cmdInput.Parameters.Add("@p1", GetSqlDbType(expectedBaseTypeName));
+                                cmdInput.Parameters["@p1"].Value = paramValue;
+                                using SqlDataReader drInput = cmdInput.ExecuteReader(CommandBehavior.CloseConnection);
+                                using SqlCommand cmd = conn.CreateCommand();
+                                cmd.CommandText = "select f1, sql_variant_property(f1,'BaseType') as BaseType from @tvpParam";
+                                SqlParameter p = cmd.Parameters.AddWithValue("@tvpParam", drInput);
+                                p.SqlDbType = SqlDbType.Structured;
+                                p.TypeName = string.Format("dbo.{0}", tvpTypeName);
+                                using (SqlDataReader dr = cmd.ExecuteReader())
+                                {
+                                    dr.Read();
+                                    return new TestResult(dr[0], dr.GetString(1));
+                                }
+                            }
+                        }
+                    }
+                    finally
+                    {
+                        DropType(conn, tvpTypeName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSqlDataReaderParameterToTVP_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSqlDataReaderParameterToTVP_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string tvpTypeName = DataTestUtility.GetLongName("tvpVariant");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropType(conn, tvpTypeName);
+                        xsql(conn, string.Format("create type dbo.{0} as table (f1 sql_variant)", tvpTypeName));
+
+                        // Send TVP using SqlDataReader.
+                        using (SqlConnection connInput = new(connStr))
+                        {
+                            connInput.Open();
+                            using (SqlCommand cmdInput = connInput.CreateCommand())
+                            {
+                                cmdInput.CommandText = "select @p1 as f1";
+                                cmdInput.Parameters.Add("@p1", SqlDbType.Variant);
+                                cmdInput.Parameters["@p1"].Value = paramValue;
+                                using SqlDataReader drInput = cmdInput.ExecuteReader(CommandBehavior.CloseConnection);
+                                using SqlCommand cmd = conn.CreateCommand();
+                                cmd.CommandText = "select f1, sql_variant_property(f1,'BaseType') as BaseType from @tvpParam";
+                                SqlParameter p = cmd.Parameters.AddWithValue("@tvpParam", drInput);
+                                p.SqlDbType = SqlDbType.Structured;
+                                p.TypeName = string.Format("dbo.{0}", tvpTypeName);
+                                using (SqlDataReader dr = cmd.ExecuteReader())
+                                {
+                                    dr.Read();
+                                    return new TestResult(dr[0], dr.GetString(1));
+                                }
+                            }
+                        }
+                    }
+                    finally
+                    {
+                        DropType(conn, tvpTypeName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSqlDataReader_TVP_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSqlDataReader_TVP_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string tvpTypeName = DataTestUtility.GetLongName("tvpType");
+                    string InputTableName = DataTestUtility.GetLongName("InputTable");
+                    string OutputTableName = DataTestUtility.GetLongName("OutputTable");
+                    string ProcName = DataTestUtility.GetLongName("spTVPProc");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+
+                        DropStoredProcedure(conn, ProcName);
+                        DropTable(conn, InputTableName);
+                        DropTable(conn, OutputTableName);
+                        DropType(conn, $"dbo.{tvpTypeName}");
+
+                        xsql(conn, string.Format("create type dbo.{0} as table (f1 {1})", tvpTypeName, expectedBaseTypeName));
+                        xsql(conn, string.Format("create table {0} (f1 {1})", InputTableName, expectedBaseTypeName));
+                        xsql(conn, string.Format("create table {0} (f1 {1})", OutputTableName, expectedBaseTypeName));
+
+                        string value = string.Empty;
+                        if (paramValue.GetType() == typeof(DateTimeOffset))
+                        {
+                            DateTime dt = ((DateTimeOffset)paramValue).UtcDateTime;
+                            value = dt.ToString("M/d/yyyy") + " " + dt.TimeOfDay;
+                        }
+                        else if (paramValue.GetType() == typeof(TimeSpan))
+                        {
+                            value = ((TimeSpan)paramValue).ToString();
+                        }
+                        else
+                        {
+                            value = ((DateTime)paramValue).ToString("M/d/yyyy") + " " + ((DateTime)paramValue).TimeOfDay;
+                        }
+                        xsql(conn, string.Format("insert into {0} values(CAST('{1}' AS {2}))", InputTableName, value, expectedBaseTypeName));
+                        xsql(conn, string.Format("create proc {0} (@P {1} READONLY) as begin insert into {2} select * from @P; end", ProcName, tvpTypeName, OutputTableName));
+
+                        SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = string.Format("SELECT * FROM {0}", InputTableName);
+                        using SqlDataReader r = cmd.ExecuteReader();
+                        using (SqlConnection conn2 = new(connStr))
+                        {
+                            conn2.Open();
+                            SqlCommand cmd2 = new(ProcName, conn2)
+                            {
+                                CommandType = CommandType.StoredProcedure
+                            };
+                            SqlParameter p = cmd2.Parameters.AddWithValue("@P", r);
+                            p.SqlDbType = SqlDbType.Structured;
+                            p.TypeName = tvpTypeName;
+                            cmd2.ExecuteNonQuery();
+
+                            cmd2.CommandText = string.Format("SELECT f1, sql_variant_property(f1,'BaseType') as BaseType FROM {0}", OutputTableName);
+                            cmd2.CommandType = CommandType.Text;
+                            using (SqlDataReader dr = cmd2.ExecuteReader())
+                            {
+                                dr.Read();
+                                return new TestResult(dr[0], dr.GetString(1));
+                            }
+                        }
+                    }
+                    finally
+                    {
+                        DropStoredProcedure(conn, ProcName);
+                        DropTable(conn, InputTableName);
+                        DropTable(conn, OutputTableName);
+                        DropType(conn, tvpTypeName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSqlDataReader_TVP_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSqlDataReader_TVP_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string tvpTypeName = DataTestUtility.GetLongName("tvpVariant_DRdrTVPVar");
+                    string InputTableName = DataTestUtility.GetLongName("InputTable");
+                    string OutputTableName = DataTestUtility.GetLongName("OutputTable");
+                    string ProcName = DataTestUtility.GetLongName("spTVPProc_DRdrTVPVar");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+
+                        DropStoredProcedure(conn, ProcName);
+                        DropTable(conn, InputTableName);
+                        DropTable(conn, OutputTableName);
+                        DropType(conn, tvpTypeName);
+
+                        xsql(conn, string.Format("create type {0} as table (f1 sql_variant)", tvpTypeName));
+                        xsql(conn, string.Format("create table {0} (f1 sql_variant)", InputTableName));
+                        xsql(conn, string.Format("create table {0} (f1 sql_variant)", OutputTableName));
+
+                        string value = string.Empty;
+                        if (paramValue.GetType() == typeof(DateTimeOffset))
+                        {
+                            DateTime dt = ((DateTimeOffset)paramValue).UtcDateTime;
+                            value = dt.ToString("M/d/yyyy") + " " + dt.TimeOfDay;
+                        }
+                        else if (paramValue.GetType() == typeof(TimeSpan))
+                        {
+                            value = ((TimeSpan)paramValue).ToString();
+                        }
+                        else
+                        {
+                            value = ((DateTime)paramValue).ToString("M/d/yyyy") + " " + ((DateTime)paramValue).TimeOfDay;
+                        }
+                        xsql(conn, string.Format("insert into {0} values(CAST('{1}' AS {2}))", InputTableName, value, expectedBaseTypeName));
+                        xsql(conn, string.Format("create proc {0} (@P {1} READONLY) as begin insert into {2} select * from @P; end", ProcName, tvpTypeName, OutputTableName));
+
+                        SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = string.Format("SELECT * FROM {0}", InputTableName);
+                        using SqlDataReader r = cmd.ExecuteReader();
+                        using (SqlConnection conn2 = new(connStr))
+                        {
+                            conn2.Open();
+                            using (SqlCommand cmd2 = new(ProcName, conn2))
+                            {
+                                cmd2.CommandType = CommandType.StoredProcedure;
+                                SqlParameter p = cmd2.Parameters.AddWithValue("@P", r);
+                                p.SqlDbType = SqlDbType.Structured;
+                                p.TypeName = tvpTypeName;
+                                cmd2.ExecuteNonQuery();
+
+                                cmd2.CommandText = string.Format("SELECT f1, sql_variant_property(f1,'BaseType') as BaseType FROM {0}", OutputTableName);
+                                cmd2.CommandType = CommandType.Text;
+                                using (SqlDataReader dr = cmd2.ExecuteReader())
+                                {
+                                    dr.Read();
+                                    return new TestResult(dr[0], dr.GetString(1));
+                                }
+                            }
+                        }
+                    }
+                    finally
+                    {
+                        DropStoredProcedure(conn, ProcName);
+                        DropTable(conn, InputTableName);
+                        DropTable(conn, OutputTableName);
+                        DropType(conn, tvpTypeName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSimpleDataReader_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSimpleDataReader_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string inputTable = DataTestUtility.GetLongName("inputTable");
+                    string procName = DataTestUtility.GetLongName("paramProc3");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, inputTable);
+                        DropStoredProcedure(conn, procName);
+
+                        string value = string.Empty;
+                        if (paramValue.GetType() == typeof(DateTimeOffset))
+                        {
+                            DateTime dt = ((DateTimeOffset)paramValue).UtcDateTime;
+                            value = dt.ToString("M/d/yyyy") + " " + dt.TimeOfDay;
+                        }
+                        else if (paramValue.GetType() == typeof(TimeSpan))
+                        {
+                            value = ((TimeSpan)paramValue).ToString();
+                        }
+                        else
+                        {
+                            value = ((DateTime)paramValue).ToString("M/d/yyyy") + " " + ((DateTime)paramValue).TimeOfDay;
+                        }
+                        xsql(conn, string.Format("create table {0} (f1 {1})", inputTable, expectedBaseTypeName));
+                        xsql(conn, string.Format("insert into {0}(f1) values('{1}');", inputTable, value));
+                        xsql(conn, string.Format("create proc {0} as begin select f1, sql_variant_property(f1,'BaseType') as BaseType from {1} end;", procName, inputTable));
+
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = procName;
+                        cmd.CommandType = CommandType.StoredProcedure;
+                        using (SqlDataReader dr = cmd.ExecuteReader())
+                        {
+                            dr.Read();
+                            return new TestResult(dr[0], dr.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropStoredProcedure(conn, procName);
+                        DropTable(conn, inputTable);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void TestSimpleDataReader_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.TestSimpleDataReader_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string inputTable = DataTestUtility.GetLongName("inputTable");
+                    string procName = DataTestUtility.GetLongName("paramProc4");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, inputTable);
+                        DropStoredProcedure(conn, procName);
+
+                        string value = string.Empty;
+                        if (paramValue.GetType() == typeof(DateTimeOffset))
+                        {
+                            DateTime dt = ((DateTimeOffset)paramValue).UtcDateTime;
+                            value = dt.ToString("M/d/yyyy") + " " + dt.TimeOfDay;
+                        }
+                        else if (paramValue.GetType() == typeof(TimeSpan))
+                        {
+                            value = ((TimeSpan)paramValue).ToString();
+                        }
+                        else
+                        {
+                            value = ((DateTime)paramValue).ToString("M/d/yyyy") + " " + ((DateTime)paramValue).TimeOfDay;
+                        }
+                        xsql(conn, string.Format("create table {0} (f1 sql_variant)", inputTable));
+                        xsql(conn, string.Format("insert into {0}(f1) values(CAST('{1}' AS {2}));", inputTable, value, expectedBaseTypeName));
+                        xsql(conn, string.Format("create proc {0} as begin select f1, sql_variant_property(f1,'BaseType') as BaseType from {1} end;", procName, inputTable));
+
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = procName;
+                        cmd.CommandType = CommandType.StoredProcedure;
+                        using (SqlDataReader dr = cmd.ExecuteReader())
+                        {
+                            dr.Read();
+                            return new TestResult(dr[0], dr.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropStoredProcedure(conn, procName);
+                        DropTable(conn, inputTable);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void SqlBulkCopySqlDataReader_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.SqlBulkCopySqlDataReader_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string bulkCopySrcTableName = DataTestUtility.GetLongName("bulkSrcTable");
+                    string bulkCopyTableName = DataTestUtility.GetLongName("bulkDestTable");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, bulkCopyTableName);
+                        xsql(conn, string.Format("create table {0} (f1 {1})", bulkCopyTableName, expectedBaseTypeName));
+
+                        DropTable(conn, bulkCopySrcTableName);
+                        xsql(conn, string.Format("create table {0} (f1 {1})", bulkCopySrcTableName, expectedBaseTypeName));
+                        string value = string.Empty;
+                        if (paramValue.GetType() == typeof(DateTimeOffset))
+                        {
+                            DateTime dt = ((DateTimeOffset)paramValue).UtcDateTime;
+                            value = dt.ToString("M/d/yyyy") + " " + dt.TimeOfDay;
+                        }
+                        else if (paramValue.GetType() == typeof(TimeSpan))
+                        {
+                            value = ((TimeSpan)paramValue).ToString();
+                        }
+                        else
+                        {
+                            value = ((DateTime)paramValue).ToString("M/d/yyyy") + " " + ((DateTime)paramValue).TimeOfDay;
+                        }
+                        xsql(conn, string.Format("insert into {0}(f1) values(CAST('{1}' AS {2}));", bulkCopySrcTableName, value, expectedBaseTypeName));
+
+                        using SqlConnection connInput = new(connStr);
+                        connInput.Open();
+                        using (SqlCommand cmdInput = connInput.CreateCommand())
+                        {
+                            cmdInput.CommandText = string.Format("select * from {0}", bulkCopySrcTableName);
+                            using SqlDataReader drInput = cmdInput.ExecuteReader();
+                            // Perform bulk copy to target.
+                            using (SqlBulkCopy bulkCopy = new(conn))
+                            {
+                                bulkCopy.BulkCopyTimeout = 60;
+                                bulkCopy.BatchSize = 1;
+                                bulkCopy.DestinationTableName = bulkCopyTableName;
+                                bulkCopy.WriteToServer(drInput);
+                            }
+
+                            // Verify target.
+                            using SqlCommand cmd = conn.CreateCommand();
+                            cmd.CommandText = string.Format("select f1, sql_variant_property(f1,'BaseType') as BaseType from {0}", bulkCopyTableName);
+                            using (SqlDataReader drVerify = cmd.ExecuteReader())
+                            {
+                                drVerify.Read();
+                                return new TestResult(drVerify[0], drVerify.GetString(1));
+                            }
+                        }
+                    }
+                    finally
+                    {
+                        DropTable(conn, bulkCopyTableName);
+                        DropTable(conn, bulkCopySrcTableName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void SqlBulkCopySqlDataReader_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.SqlBulkCopySqlDataReader_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string bulkCopySrcTableName = DataTestUtility.GetLongName("bulkSrcTable");
+                    string bulkCopyTableName = DataTestUtility.GetLongName("bulkDestTable");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, bulkCopyTableName);
+                        xsql(conn, string.Format("create table {0} (f1 sql_variant)", bulkCopyTableName));
+
+                        DropTable(conn, bulkCopySrcTableName);
+                        xsql(conn, string.Format("create table {0} (f1 {1})", bulkCopySrcTableName, expectedBaseTypeName));
+                        string value = string.Empty;
+                        if (paramValue.GetType() == typeof(DateTimeOffset))
+                        {
+                            DateTime dt = ((DateTimeOffset)paramValue).UtcDateTime;
+                            value = dt.ToString("M/d/yyyy") + " " + dt.TimeOfDay;
+                        }
+                        else if (paramValue.GetType() == typeof(TimeSpan))
+                        {
+                            value = ((TimeSpan)paramValue).ToString();
+                        }
+                        else
+                        {
+                            value = ((DateTime)paramValue).ToString("M/d/yyyy") + " " + ((DateTime)paramValue).TimeOfDay;
+                        }
+                        xsql(conn, string.Format("insert into {0}(f1) values(CAST('{1}' AS {2}));", bulkCopySrcTableName, value, expectedBaseTypeName));
+
+                        using (SqlConnection connInput = new(connStr))
+                        {
+                            connInput.Open();
+                            using (SqlCommand cmdInput = connInput.CreateCommand())
+                            {
+                                cmdInput.CommandText = string.Format("select * from {0}", bulkCopySrcTableName);
+                                using SqlDataReader drInput = cmdInput.ExecuteReader();
+                                {
+                                    // Perform bulk copy to target.
+                                    using (SqlBulkCopy bulkCopy = new(conn))
+                                    {
+                                        bulkCopy.BulkCopyTimeout = 60;
+                                        bulkCopy.BatchSize = 1;
+                                        bulkCopy.DestinationTableName = bulkCopyTableName;
+                                        bulkCopy.WriteToServer(drInput);
+                                    }
+
+                                    // Verify target.
+                                    using SqlCommand cmd = conn.CreateCommand();
+                                    cmd.CommandText = string.Format("select f1, sql_variant_property(f1,'BaseType') as BaseType from {0}", bulkCopyTableName);
+                                    using (SqlDataReader drVerify = cmd.ExecuteReader())
+                                    {
+                                        drVerify.Read();
+                                        return new TestResult(drVerify[0], drVerify.GetString(1));
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    finally
+                    {
+                        DropTable(conn, bulkCopyTableName);
+                        DropTable(conn, bulkCopySrcTableName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void SqlBulkCopyDataTable_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.SqlBulkCopyDataTable_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string bulkCopyTableName = DataTestUtility.GetLongName("bulkDestType");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, bulkCopyTableName);
+                        xsql(conn, string.Format("create table {0} (f1 {1})", bulkCopyTableName, expectedBaseTypeName));
+
+                        // Send using DataTable as source.
+                        DataTable t = new();
+                        t.Columns.Add("f1", paramValue.GetType());
+                        t.Rows.Add(new object[] { paramValue });
+
+                        // Perform bulk copy to target.
+                        using (SqlBulkCopy bulkCopy = new(conn))
+                        {
+                            bulkCopy.BulkCopyTimeout = 60;
+                            bulkCopy.BatchSize = 1;
+                            bulkCopy.DestinationTableName = bulkCopyTableName;
+                            bulkCopy.WriteToServer(t, DataRowState.Added);
+                        }
+
+                        // Verify target.
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = string.Format("select f1, sql_variant_property(f1,'BaseType') as BaseType from {0}", bulkCopyTableName);
+                        using (SqlDataReader drVerify = cmd.ExecuteReader())
+                        {
+                            drVerify.Read();
+                            return new TestResult(drVerify[0], drVerify.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropTable(conn, bulkCopyTableName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void SqlBulkCopyDataTable_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.SqlBulkCopyDataTable_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string bulkCopyTableName = DataTestUtility.GetLongName("bulkDestVariant");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, bulkCopyTableName);
+                        xsql(conn, string.Format("create table {0} (f1 sql_variant)", bulkCopyTableName));
+
+                        // Send using DataTable as source.
+                        DataTable t = new();
+                        t.Columns.Add("f1", typeof(object));
+                        t.Rows.Add(new object[] { paramValue });
+
+                        // Perform bulk copy to target.
+                        using (SqlBulkCopy bulkCopy = new(conn))
+                        {
+                            bulkCopy.BulkCopyTimeout = 60;
+                            bulkCopy.BatchSize = 1;
+                            bulkCopy.DestinationTableName = bulkCopyTableName;
+                            bulkCopy.WriteToServer(t, DataRowState.Added);
+                        }
+
+                        // Verify target.
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = string.Format("select f1, sql_variant_property(f1,'BaseType') as BaseType from {0}", bulkCopyTableName);
+                        using (SqlDataReader drVerify = cmd.ExecuteReader())
+                        {
+                            drVerify.Read();
+                            return new TestResult(drVerify[0], drVerify.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropTable(conn, bulkCopyTableName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void SqlBulkCopyDataRow_Type(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.SqlBulkCopyDataRow_Type,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string bulkCopyTableName = DataTestUtility.GetLongName("bulkDestType");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, bulkCopyTableName);
+                        xsql(conn, string.Format("create table {0} (f1 {1})", bulkCopyTableName, expectedBaseTypeName));
+                        DataTable t = new();
+                        t.Columns.Add("f1", paramValue.GetType());
+                        t.Rows.Add(new object[] { paramValue });
+                        DataRow[] rowToSend = t.Select();
+                        using (SqlBulkCopy bulkCopy = new(conn))
+                        {
+                            bulkCopy.BulkCopyTimeout = 60;
+                            bulkCopy.BatchSize = 1;
+                            bulkCopy.DestinationTableName = bulkCopyTableName;
+                            bulkCopy.WriteToServer(rowToSend);
+                        }
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = string.Format("select f1, sql_variant_property(f1,'BaseType') as BaseType from {0}", bulkCopyTableName);
+                        using (SqlDataReader drVerify = cmd.ExecuteReader())
+                        {
+                            drVerify.Read();
+                            return new TestResult(drVerify[0], drVerify.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropTable(conn, bulkCopyTableName);
+                    }
+                },
+                paramValue, 
+                expectedBaseTypeName, 
+                DataTestUtility.TCPConnectionString, 
+                expectedExceptions, 
+                expectedValueOverrides,
+                expectedBaseTypeOverrides);
+        }
+
+        [ConditionalTheory(typeof(DataTestUtility), nameof(DataTestUtility.AreConnStringsSetup), nameof(DataTestUtility.IsNotAzureSynapse))]
+        [MemberData(nameof(GetParameterCombinations))]
+        public void SqlBulkCopyDataRow_Variant(
+            object paramValue, 
+            string expectedBaseTypeName, 
+            Dictionary<TestVariations, ExceptionChecker> expectedExceptions, 
+            Dictionary<TestVariations, object> expectedValueOverrides,
+            Dictionary<TestVariations, string> expectedBaseTypeOverrides)
+        {
+            RunTest(
+                TestVariations.SqlBulkCopyDataRow_Variant,
+                (paramValue, expectedBaseTypeName, connStr) =>
+                {
+                    string bulkCopyTableName = DataTestUtility.GetLongName("bulkDestVariant");
+
+                    using SqlConnection conn = new(connStr);
+                    try
+                    {
+                        conn.Open();
+                        DropTable(conn, bulkCopyTableName);
+                        xsql(conn, string.Format("create table {0} (f1 sql_variant)", bulkCopyTableName));
+                        DataTable t = new();
+                        t.Columns.Add("f1", typeof(object));
+                        t.Rows.Add(new object[] { paramValue });
+                        DataRow[] rowToSend = t.Select();
+                        using (SqlBulkCopy bulkCopy = new(conn))
+                        {
+                            bulkCopy.BulkCopyTimeout = 60;
+                            bulkCopy.BatchSize = 1;
+                            bulkCopy.DestinationTableName = bulkCopyTableName;
+                            bulkCopy.WriteToServer(rowToSend);
+                        }
+                        using SqlCommand cmd = conn.CreateCommand();
+                        cmd.CommandText = string.Format("select f1, sql_variant_property(f1,'BaseType') as BaseType from {0}", bulkCopyTableName);
+                        using (SqlDataReader drVerify = cmd.ExecuteReader())
+                        {
+                            drVerify.Read();
+                            return new TestResult(drVerify[0], drVerify.GetString(1));
+                        }
+                    }
+                    finally
+                    {
+                        DropTable(conn, bulkCopyTableName);
+                    }
+                },
+                paramValue, 
                 expectedBaseTypeName, 
                 DataTestUtility.TCPConnectionString, 
                 expectedExceptions, 
@@ -51,19 +1125,19 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
         /// Gets parameter combinations as indices for MemberData.
         /// Using indices for xUnit serialization compatibility.
         /// </summary>
-        public static IEnumerable<object?[]> GetParameterCombinations()
+        public static IEnumerable<object[]> GetParameterCombinations()
         {
-            yield return new object?[] { DateTime.MinValue, "System.DateTime", "date",
+            yield return new object[] { DateTime.MinValue, "date",
                 new Dictionary<TestVariations, ExceptionChecker> {
                     { TestVariations.TestSimpleParameter_Variant, SqlDateTimeOverflow },
                     { TestVariations.TestSqlDataRecordParameterToTVP_Variant, SqlDateTimeOverflow },
                     { TestVariations.TestSqlDataReaderParameterToTVP_Variant, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataTable_Variant, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Variant, SqlDateTimeOverflow }}, 
-                null, 
-                null};
-            yield return new object?[] { DateTime.MaxValue, "System.DateTime", "date", 
-                null, 
+                new Dictionary<TestVariations, object>(), 
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.MaxValue, "date", 
+                new Dictionary<TestVariations, ExceptionChecker>(), 
                 new Dictionary<TestVariations, object>
                 {
                     { TestVariations.TestSimpleParameter_Type, new DateTime(3155378112000000000) },
@@ -83,7 +1157,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopyDataRow_Type, new DateTime(3155378112000000000) },
                     { TestVariations.SqlBulkCopyDataRow_Variant, new DateTime(3155378975999970000) }
                 },
-                new Dictionary<TestVariations, object>
+                new Dictionary<TestVariations, string>
                 {
                     {TestVariations.TestSimpleParameter_Variant, "datetime"},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, "datetime"},
@@ -92,17 +1166,17 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     {TestVariations.SqlBulkCopyDataRow_Variant, "datetime"}
                 }
             };
-            yield return new object?[] { DateTime.MinValue, "System.DateTime", "datetime2",
+            yield return new object[] { DateTime.MinValue, "datetime2",
                 new Dictionary<TestVariations, ExceptionChecker> {
                     { TestVariations.TestSimpleParameter_Variant, SqlDateTimeOverflow },
                     { TestVariations.TestSqlDataRecordParameterToTVP_Variant, SqlDateTimeOverflow },
                     { TestVariations.TestSqlDataReaderParameterToTVP_Variant, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataTable_Variant, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Variant, SqlDateTimeOverflow }}, 
-                null,
-                null};
-            yield return new object?[] { DateTime.MaxValue, "System.DateTime", "datetime2", 
-                null,
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.MaxValue, "datetime2", 
+                new Dictionary<TestVariations, ExceptionChecker>(),
                 new Dictionary<TestVariations, object> {
                     {TestVariations.TestSimpleParameter_Variant, new DateTime(3155378975999970000)},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, new DateTime(3155378975999970000)},
@@ -110,7 +1184,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     {TestVariations.SqlBulkCopyDataTable_Variant, new DateTime(3155378975999970000)},
                     {TestVariations.SqlBulkCopyDataRow_Variant, new DateTime(3155378975999970000)}
                 },
-                new Dictionary<TestVariations, object>
+                new Dictionary<TestVariations, string>
                 {
                     {TestVariations.TestSimpleParameter_Variant, "datetime"},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, "datetime"},
@@ -118,7 +1192,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     {TestVariations.SqlBulkCopyDataTable_Variant, "datetime"},
                     {TestVariations.SqlBulkCopyDataRow_Variant, "datetime"}
                 }};
-            yield return new object?[] { DateTime.MinValue, "System.DateTime", "datetime", 
+            yield return new object[] { DateTime.MinValue, "datetime", 
                 new Dictionary<TestVariations, ExceptionChecker> { 
                     { TestVariations.TestSimpleParameter_Type, SqlDateTimeOverflow },
                     { TestVariations.TestSimpleParameter_Variant, SqlDateTimeOverflow },
@@ -136,9 +1210,9 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopyDataTable_Variant, SqlDateTimeOverflow},
                     { TestVariations.SqlBulkCopyDataRow_Type, SqlDateTimeOverflow},
                     { TestVariations.SqlBulkCopyDataRow_Variant, SqlDateTimeOverflow}},
-                null,
-                null};
-            yield return new object?[] { DateTime.MaxValue, "System.DateTime", "datetime", 
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.MaxValue, "datetime", 
                 new Dictionary<TestVariations, ExceptionChecker> { 
                     { TestVariations.TestSqlDataReader_TVP_Type, CannotConvertCharacterStringToDateOrTime},
                     { TestVariations.TestSqlDataReader_TVP_Variant, CannotConvertCharacterStringToDateOrTime},
@@ -165,12 +1239,21 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopyDataRow_Type, new DateTime(3155378975999970000) },
                     { TestVariations.SqlBulkCopyDataRow_Variant, new DateTime(3155378975999970000) }
                 },
-                null};
-            yield return new object[] { DateTimeOffset.MinValue, "System.DateTimeOffset", "datetimeoffset"};
-            yield return new object[] { DateTimeOffset.MaxValue, "System.DateTimeOffset", "datetimeoffset"};
-            yield return new object[] { DateTimeOffset.Parse("12/31/1999 23:59:59.9999999 -08:30"), "System.DateTimeOffset", "datetimeoffset"};
-            yield return new object?[] { DateTime.Parse("1998-01-01 23:59:59.995"), "System.DateTime", "datetime2",
-                null,
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTimeOffset.MinValue, "datetimeoffset",
+                new Dictionary<TestVariations, ExceptionChecker>(),
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTimeOffset.MaxValue, "datetimeoffset",
+                new Dictionary<TestVariations, ExceptionChecker>(),
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTimeOffset.Parse("12/31/1999 23:59:59.9999999 -08:30"), "datetimeoffset",
+                new Dictionary<TestVariations, ExceptionChecker>(),
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.Parse("1998-01-01 23:59:59.995"), "datetime2",
+                new Dictionary<TestVariations, ExceptionChecker>(),
                 new Dictionary<TestVariations, object>
                 {
                     {TestVariations.TestSimpleParameter_Variant, DateTime.Parse("1998-01-01 23:59:59.997")},
@@ -179,7 +1262,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     {TestVariations.SqlBulkCopyDataTable_Variant, DateTime.Parse("1998-01-01 23:59:59.997")},
                     {TestVariations.SqlBulkCopyDataRow_Variant, DateTime.Parse("1998-01-01 23:59:59.997")}
                 },
-                new Dictionary<TestVariations, object>
+                new Dictionary<TestVariations, string>
                 {
                     {TestVariations.TestSimpleParameter_Variant, "datetime"},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, "datetime"},
@@ -187,7 +1270,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     {TestVariations.SqlBulkCopyDataTable_Variant, "datetime"},
                     {TestVariations.SqlBulkCopyDataRow_Variant, "datetime"}
                 }};
-            yield return new object?[] { DateTime.MinValue, "System.DateTime", "smalldatetime",
+            yield return new object[] { DateTime.MinValue, "smalldatetime",
                 new Dictionary<TestVariations, ExceptionChecker> {
                     { TestVariations.TestSimpleParameter_Type, SqlDateTimeOverflow },
                     { TestVariations.TestSimpleParameter_Variant, SqlDateTimeOverflow },
@@ -205,9 +1288,9 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopyDataTable_Variant, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Type, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Variant, SqlDateTimeOverflow }},
-                null,
-                null};
-            yield return new object?[] { DateTime.MaxValue, "System.DateTime", "smalldatetime",
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.MaxValue, "smalldatetime",
                 new Dictionary<TestVariations, ExceptionChecker> {
                     { TestVariations.TestSimpleParameter_Type, UnRepresentableDateTime },
                     { TestVariations.TestSqlDataRecordParameterToTVP_Type, InvalidValueForMetadata },
@@ -227,7 +1310,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopyDataTable_Variant, new DateTime(3155378975999970000) },
                     { TestVariations.SqlBulkCopyDataRow_Variant, new DateTime(3155378975999970000) }
                 },
-                new Dictionary<TestVariations, object>
+                new Dictionary<TestVariations, string>
                 {
                     {TestVariations.TestSimpleParameter_Variant, "datetime"},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, "datetime"},
@@ -235,7 +1318,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     {TestVariations.SqlBulkCopyDataTable_Variant, "datetime"},
                     {TestVariations.SqlBulkCopyDataRow_Variant, "datetime"}
                 }};
-            yield return new object?[] { TimeSpan.MinValue, "System.TimeSpan", "time",
+            yield return new object[] { TimeSpan.MinValue, "time",
                 new Dictionary<TestVariations, ExceptionChecker> {
                     { TestVariations.TestSimpleParameter_Type, TimeOverflow },
                     { TestVariations.TestSimpleParameter_Variant, TimeOverflow },
@@ -255,8 +1338,8 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                 new Dictionary<TestVariations, object> {
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, TimeSpan.Zero},
                 },
-                null};
-            yield return new object?[] { TimeSpan.MaxValue, "System.TimeSpan", "time",
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { TimeSpan.MaxValue, "time",
                 new Dictionary<TestVariations, ExceptionChecker> {
                     { TestVariations.TestSimpleParameter_Type, TimeOverflow },
                     { TestVariations.TestSimpleParameter_Variant, TimeOverflow },
@@ -274,9 +1357,9 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopyDataTable_Variant, TimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Type, TimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Variant, TimeOverflow }},
-                null,
-                null};
-            yield return new object?[] { DateTime.MinValue, "System.DateTime", "time",
+                new Dictionary<TestVariations, object>(),
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.MinValue, "time",
                 new Dictionary<TestVariations, ExceptionChecker> { 
                     { TestVariations.SqlBulkCopyDataTable_Type, CannotConvertMinDateTimeToTime},
                     { TestVariations.SqlBulkCopyDataRow_Type, CannotConvertMinDateTimeToTime},
@@ -294,9 +1377,16 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopySqlDataReader_Variant, InvalidCastNotValid },
                     { TestVariations.SqlBulkCopyDataTable_Variant, SqlDateTimeOverflow },
                     { TestVariations.SqlBulkCopyDataRow_Variant, SqlDateTimeOverflow }}, 
-                null,
-                null};
-            yield return new object?[] { DateTime.MaxValue, "System.DateTime", "time",
+                new Dictionary<TestVariations, object> {
+                    {TestVariations.SqlBulkCopySqlDataReader_Type, TimeSpan.Zero},
+                    {TestVariations.SqlBulkCopySqlDataReader_Variant, TimeSpan.Zero},
+                    {TestVariations.TestSqlDataReader_TVP_Type, TimeSpan.Zero},
+                    {TestVariations.TestSqlDataReader_TVP_Variant, TimeSpan.Zero},
+                    {TestVariations.TestSimpleDataReader_Type, TimeSpan.Zero},
+                    {TestVariations.TestSimpleDataReader_Variant, TimeSpan.Zero},
+                },
+                new Dictionary<TestVariations, string>()};
+            yield return new object[] { DateTime.MaxValue, "time",
                 new Dictionary<TestVariations, ExceptionChecker> { 
                     { TestVariations.SqlBulkCopyDataTable_Type, CannotConvertMaxDateTimeToTime },
                     { TestVariations.SqlBulkCopyDataRow_Type, CannotConvertMaxDateTimeToTime },
@@ -310,13 +1400,19 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                     { TestVariations.SqlBulkCopySqlDataReader_Type, InvalidCastNotValid },
                     { TestVariations.SqlBulkCopySqlDataReader_Variant, InvalidCastNotValid }}, 
                 new Dictionary<TestVariations, object> {
+                    {TestVariations.TestSqlDataReader_TVP_Variant, TimeSpan.Parse("23:59:59.9999999")},
+                    {TestVariations.SqlBulkCopySqlDataReader_Type, TimeSpan.Parse("23:59:59.9999999")},
+                    {TestVariations.SqlBulkCopySqlDataReader_Variant, TimeSpan.Parse("23:59:59.9999999")},
+                    {TestVariations.TestSqlDataReader_TVP_Type, TimeSpan.Parse("23:59:59.9999999")},
+                    {TestVariations.TestSimpleDataReader_Type, TimeSpan.Parse("23:59:59.9999999")},
+                    {TestVariations.TestSimpleDataReader_Variant, TimeSpan.Parse("23:59:59.9999999")},
                     {TestVariations.TestSimpleParameter_Variant, new DateTime(3155378975999970000)},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, new DateTime(3155378975999970000)},
                     {TestVariations.TestSqlDataReaderParameterToTVP_Variant, new DateTime(3155378975999970000)},
                     {TestVariations.SqlBulkCopyDataTable_Variant, new DateTime(3155378975999970000)},
                     {TestVariations.SqlBulkCopyDataRow_Variant, new DateTime(3155378975999970000)}
                 },
-                new Dictionary<TestVariations, object>
+                new Dictionary<TestVariations, string>
                 {
                     {TestVariations.TestSimpleParameter_Variant, "datetime"},
                     {TestVariations.TestSqlDataRecordParameterToTVP_Variant, "datetime"},
@@ -378,64 +1474,5 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             (e.GetType() == typeof(SqlException)) &&
             e.Message.Contains("The incoming tabular data stream (TDS) remote procedure call (RPC) protocol stream is incorrect.");
 
-        /// <summary>
-        /// Gets the baseline file for a specific parameter index and compares against output.
-        /// Each parameter has a single combined baseline file: DateTimeVariant_XX.bsl
-        /// </summary>
-        private static string FindDiffFromBaselineFiles(int paramIndex, string outputPath)
-        {
-            // Get the baseline file for this parameter
-            string baselineFile = Path.Combine(BaselineDirectory, $"DateTimeVariant_{paramIndex:D2}.bsl");
-
-            if (!File.Exists(baselineFile))
-            {
-                return $"Baseline file not found: {baselineFile}";
-            }
-
-            var expectedLines = File.ReadAllLines(baselineFile);
-
-            var outputLines = File.ReadAllLines(outputPath);
-
-            var comparisonSb = new StringBuilder();
-
-            var expectedLength = expectedLines.Length;
-            var outputLength = outputLines.Length;
-            var findDiffLength = Math.Min(expectedLength, outputLength);
-
-            for (var lineNo = 0; lineNo < findDiffLength; lineNo++)
-            {
-                if (!expectedLines[lineNo].Equals(outputLines[lineNo]))
-                {
-                    comparisonSb.AppendFormat("** DIFF at line {0} \n", lineNo);
-                    comparisonSb.AppendFormat("A : {0} \n", outputLines[lineNo]);
-                    comparisonSb.AppendFormat("E : {0} \n", expectedLines[lineNo]);
-                }
-            }
-
-            var startIndex = findDiffLength - 1;
-            if (startIndex < 0)
-            {
-                startIndex = 0;
-            }
-
-            if (findDiffLength < expectedLength)
-            {
-                comparisonSb.AppendFormat("** MISSING \n");
-                for (var lineNo = startIndex; lineNo < expectedLength; lineNo++)
-                {
-                    comparisonSb.AppendFormat("{0} : {1}", lineNo, expectedLines[lineNo]);
-                }
-            }
-            if (findDiffLength < outputLength)
-            {
-                comparisonSb.AppendFormat("** EXTRA \n");
-                for (var lineNo = startIndex; lineNo < outputLength; lineNo++)
-                {
-                    comparisonSb.AppendFormat("{0} : {1}", lineNo, outputLines[lineNo]);
-                }
-            }
-
-            return comparisonSb.ToString();
-        }
     }
 }
